library refactor
